+++ vortex-ipc/src/stream_reader/mod.rs
